feat(theme): Upgrade agnoster theme
My old version of the agnoster theme, in which I had altered:

  * Colours used
  * Prompt segments

I couldn't make the customizing of the theme work, so I have fallen
back on my old way of copying the agnoster theme to a personal copy
and edited it directly. It would be preferrable to configure it though.