fix(environment): Remove obsolete settings and tidy up
This is a big change, that really should have been done in separate
steps, but I've not got the time.

I have removed commented out and obsolete settings in the PATH, exports,
aliases etc.

I have moved secrets that are exported so apps can work into a separate
file that is sourced from here and ignored, so I cannot accidentally
commit those secrets.